+++ resources/js/app.js
@@ -26,6 +26,7 @@ import MainView from './components/MainView.vue';
 import Login from './components/Login.vue';
 import ConceptDetail from './components/ConceptDetail.vue';
 import ErrorModal from './components/ErrorModal.vue';
+import ImportingInfoModal from './components/modals/ImportingInfoModal.vue';
 import AboutDialog from './components/About.vue';
 // Settings
 import Users from './components/Users.vue';
@@ -71,7 +72,9 @@ Vue.use(VueRouter);
 Vue.use(VueI18n);
 Vue.use(VeeValidate);
 Vue.use(Notifications);
-Vue.use(VModal, {dynamic: true});
+Vue.use(VModal, {
+    dynamic: true,
+});
 Vue.use(VueScrollTo);
 
 const router = new VueRouter({
@@ -221,6 +224,7 @@ Vue.use(require('@websanova/vue-auth'), {
 // Imported Components
 Vue.component('multiselect', Multiselect);
 // Pages
+Vue.component('importing-info-modal', ImportingInfoModal);
 Vue.component('error-modal', ErrorModal);
 Vue.component('about-dialog', AboutDialog);