+++ resources/js/i18n/en.js
@@ -138,6 +138,10 @@ const en = {
             ).`,
             delete: 'Delete <span class="font-weight-medium">{name}</span>',
             'delete-recursive': 'Delete recursive'
+        },
+        import_info: {
+            title: 'Importing…',
+            info: 'Please wait while ThesauRex is importing your file. This window closes after the import is finished.'
         }
     },
     menus: {